Restructure repo and separate IO and MPI
